Do not downcase entry name
The page_entries_info helper was updated in #694 to support for internationalization of entry_name for ActiveRecord. However, it always calls downcase on the resulting entry name. This is a problem for languages like German, where the models' names (and all nouns) are capitalized. Removing the call to downcase will probably cause no problems for other languages, where translated entity names are lowercase or case does not apply.

Fixes #850.